refactor: split events.ts handler into per-event-type modules
- Extract handleSessionError into handle-session-error.ts
- Extract handleSessionCompacted into handle-session-compacted.ts
- Extract handleSessionIdle into handle-session-idle.ts
- Extract handleSessionStatus into handle-session-status.ts
- Extract handleSessionDeleted into handle-session-deleted.ts
- Simplify events.ts to dispatcher-only with re-exports

Ultraworked with [Sisyphus](https://github.com/code-yeongyu/oh-my-openagent)
